{"schema":"decodex/commit/1","summary":"Modularize frozen selection chrome renderer","authority":"manual"}
